chore(consensus): move test-context out of consensus crate
