+++ Cargo.toml
@@ -34,6 +34,7 @@ members = [
   "crates/papyrus_sync",
   "crates/papyrus_test_utils",
   "crates/sequencing/papyrus_consensus",
+  "crates/sequencing/papyrus_consensus_orchestrator",
   "crates/starknet_api",
   "crates/starknet_client",
   "crates/starknet_committer",
@@ -137,6 +138,7 @@ papyrus_base_layer = { path = "crates/papyrus_base_layer", version = "0.0.0" }
 papyrus_common = { path = "crates/papyrus_common", version = "0.0.0" }
 papyrus_config = { path = "crates/papyrus_config", version = "0.0.0" }
 papyrus_consensus = { path = "crates/sequencing/papyrus_consensus", version = "0.0.0" }
+papyrus_consensus_orchestrator = { path = "crates/sequencing/papyrus_consensus_orchestrator", version = "0.0.0" }
 papyrus_execution = { path = "crates/papyrus_execution", version = "0.0.0" }
 papyrus_monitoring_gateway = { path = "crates/papyrus_monitoring_gateway", version = "0.0.0" }
 papyrus_network = { path = "crates/papyrus_network", version = "0.0.0" }
