+++ crates/papyrus_protobuf/src/consensus.rs
@@ -1,4 +1,5 @@
-use starknet_api::block::BlockHash;
+use futures::channel::{mpsc, oneshot};
+use starknet_api::block::{BlockHash, BlockNumber};
 use starknet_api::core::ContractAddress;
 use starknet_api::transaction::Transaction;
 
@@ -49,4 +50,31 @@ pub struct StreamMessage<T: Into<Vec<u8>> + TryFrom<Vec<u8>, Error = ProtobufCon
     pub stream_id: u64,
     pub chunk_id: u64,
     pub fin: bool,
+}
+
+// TODO(Guy): Remove after implementing broadcast streams.
+#[allow(missing_docs)]
+pub struct ProposalWrapper(pub Proposal);
+
+impl From<ProposalWrapper>
+    for (
+        (BlockNumber, u32, ContractAddress),
+        mpsc::Receiver<Transaction>,
+        oneshot::Receiver<BlockHash>,
+    )
+{
+    fn from(val: ProposalWrapper) -> Self {
+        let transactions: Vec<Transaction> = val.0.transactions.into_iter().collect();
+        let proposal_init = (BlockNumber(val.0.height), val.0.round, val.0.proposer);
+        let (mut content_sender, content_receiver) = mpsc::channel(transactions.len());
+        for tx in transactions {
+            content_sender.try_send(tx).expect("Send should succeed");
+        }
+        content_sender.close_channel();
+
+        let (fin_sender, fin_receiver) = oneshot::channel();
+        fin_sender.send(val.0.block_hash).expect("Send should succeed");
+
+        (proposal_init, content_receiver, fin_receiver)
+    }
 }
