+++ crates/sequencing/papyrus_consensus_orchestrator/src/lib.rs
@@ -1,0 +1,7 @@
+#![warn(missing_docs)]
+//! An orchestrator for a StarkNet node.
+//! Implements the consensus context - the interface for consensus to call out to the node.
+
+#[allow(missing_docs)]
+// TODO: this is test code, rename accordingly.
+pub mod papyrus_consensus_context;
